Move zoom, copyPaste, reloadChanged, title to main
