+++ src/main/nvim/nvim.js
@@ -1,15 +1,21 @@
 import { ipcMain } from 'electron';
 
-import initQuit from './features/quit';
+import {
+  getNvimByWindow,
+  setNvimByWindow,
+  deleteNvimByWindow,
+} from './nvimByWindow';
+
+import quit from './features/quit';
+import windowTitle from './features/windowTitle';
+import zoom from './features/zoom';
+import reloadChanged from './features/reloadChanged';
 
 import nvimApi from './api';
 
-const nvimByWindowId = [];
-
 ipcMain.on('nvim-send', ({ sender: { id } }, payload) => {
-  if (nvimByWindowId[id]) {
-    nvimByWindowId[id].send(...payload);
-  }
+  const nvim = getNvimByWindow(id);
+  if (nvim) nvim.send(...payload);
 });
 
 const initNvim = ({ args, cwd, win }) => {
@@ -22,12 +28,15 @@ const initNvim = ({ args, cwd, win }) => {
   nvim.on('data', data => win.webContents.send('nvim-data', data));
 
   nvim.on('disconnect', () => {
-    delete nvimByWindowId[win.webContents.id];
+    deleteNvimByWindow(win);
   });
 
-  nvimByWindowId[win.webContents.id] = nvim;
+  setNvimByWindow(win, nvim);
 
-  initQuit({ win, nvim });
+  quit({ win, nvim });
+  windowTitle({ win, nvim });
+  zoom({ win });
+  reloadChanged({ win, nvim });
 };
 
 export default initNvim;
